Add input validator
