+++ controllers/user-controllers/user-controller.js
@@ -3,18 +3,18 @@
 const { createUser, userDataIsValid } = require('../../services/user-services/user-service');
 
 const createNewUser = async (req, res) => {
-  if (!userDataIsValid()) {
+  if (!userDataIsValid(req.body)) {
     return res.sendError(400, 'Invalid input data. Please check the documentation.');
   }
 
   const user = await createUser(req.body)
     .catch((error) => {
       console.log(error);
-      res.sendError(500, `Failed to create user: ${error.detail}`);
+      res.sendError(500, 'The server encountered an error.');
     });
   if (!user) return false;
 
-  res.sendData(201, user);
+  return res.sendData(201, user);
 };
 
 module.exports = createNewUser;
